+++ example/02_test/test00.c
@@ -41,6 +41,9 @@ int main(int argc, char* argv)
     vgs_putlog("vgs_strstr(\"%s\",\"is t\") ... %s", text, expect(vgs_strstr(text, "is t"), "is test string."));
     vgs_putlog("vgs_strcmp(\"AbCd\",\"aBcD\") = %d", expect32d(vgs_strcmp("AbCd", "aBcD"), -1));
     vgs_putlog("vgs_stricmp(\"AbCd\",\"aBcD\") = %d", expect32d(vgs_stricmp("AbCd", "aBcD"), 0));
+    vgs_putlog("vgs_atoi(\"12345\") = %d", expect32d(vgs_atoi("12345"), 12345));
+    vgs_putlog("vgs_atoi(\"-12345\") = %d", expect32d(vgs_atoi("-12345"), -12345));
+    vgs_putlog("vgs_atoi(\" 12345\") = %d", expect32d(vgs_atoi(" 12345"), 0));
 
     vgs_putlog("vgs_isdigit(\'0\') ... %d", expect32d(vgs_isdigit('0'), TRUE));
     vgs_putlog("vgs_isdigit(\'9\') ... %d", expect32d(vgs_isdigit('9'), TRUE));