+++ CHANGES.md
@@ -3,7 +3,8 @@
 ## Version 0.5.0
 
 - CRT: Added the i-math inline functions: `vgs_abs` and `vgs_sgn`
-- CRT: Added a string function: `stricmp`
+- CRT: Added a string function: `vgs_stricmp`
+- CRT: Added a ctype function: `vgs_atoi`
 
 ## Version 0.4.0